Tables: Make CommandContext generic + TableSchema
This is a refactor for the tables, work it includes three commits below.

This Commit refactors the CommandContext to be generic so it can
hold the SchemaObject the command is running against.

NOTE: this commit was made without the offline build profile enabled,
this is prob fixed in the next commit

This branch was build against `ajm/tables-chunk-1`.
Original commits to the `ajm/tables` branch below.

commit 7d7f82c501046cad26028cd63ce33d57e6557a2b
Author: Aaron Morton <aaron.morton@datastax.com>
Date:   Mon Jul 8 08:56:37 2024 +1200

    Made the CommentContext generic

    compiles, all unit tests pass other than session tests

commit 8efef2bf2b97cfdd7ccf9bdf912e2b7fbf65e738
Author: Aaron Morton <aaron.morton@datastax.com>
Date:   Sun Jul 7 14:18:04 2024 +1200

    WIP - DOES NOT COMPILE

    refactor the CollectionSettings into a SchemaObject

    Before making the CommandContext generic

commit 4fd3e1804433a9b1de58c702a73abfad1098d923
Author: Aaron Morton <aaron.morton@datastax.com>
Date:   Sun Jul 7 13:15:15 2024 +1200

    cleanup CommandContext