POC for table filters with codecs for findOne
This PR refactors how we process filters by moving the base class
FilterableResolver that was used by command resolvers to be
standalone class FilterResolver that has CollectionFilterResolver and
TableFilterResolver subclasses. The changes to the resolvers are mostly
to handle this change, other than findOne.

The next part is findOne updated to process the filter for a table. Using
the TableFilterResolver to make filters, which then use the
NativeTypeTableFilter and JSONCodec to map from the types used for JSOn
into what CQL expects.

This is POC work, we then need to expand this to handle all the situations
we expect.