Updated SchemaValidatable
this was the ValidatableCommandClause in the first POC

Updates to be general for validating objects against schema, and
added tests.

Some refactoring of the SchemaObject to add accessors for some fields
so they can be properly mocked.

See comments on the SchemaValidatable - is it designed to add an easy pattern
for whenever we are validating an objects