Refactor CQLSessionCache and CqlCredentials
needed a way to get a session using only the
(String tenantId, String authToken) and the way the cache key worked made
they very hard / impossible. Also the code needed a refactor.

Needed the new access technique for the OperationAttempt where we are
trying to stop carrying the DataApiRequestInfo everywhere as it's
IOC pattern mixes application and requests state (and generally to
remove the IOC injection)