remove http override for unknown server error
The base for new Driver error handling returns a server unknown error
and the config was turning this into a HTTP 500. I think may have been there
just to show we can do that, and we will want to for some errors.

But we should be returning the basic 200 and the error in the response.