Fix #2048 refactor CqlSessionCache & add callbacks
refactor the session cache to split out responsibilities, and
add unit tests for the parts. Now has the cache, credentials
creation, and session creation in different classes.

Add a callback mechanism that is used when a session is
removed from the cache, so we can remove metrics that are no
longer needed. Also the process now used for managing the
schema cache.