Fixes #2056 stop publishing metric histograms
This PR ensures that we only publish  percentiles for distribution
metrics, such as the p98 etc. And prevent publishing histograms
which can be 70 lines long.

NOTE: this means the _buckets metric is NOT PUBLISHED.

If we think a metric is per tenant then we publish a limited
number of percentiles because there will be a lot of tenants.
For all tenant metrics we publish more percentiles.

See tests in MicrometerConfigurationTests that show the metrics we
publish.